fix more-posts component
