+++ app/[locale]/components/show-more-logic.tsx
@@ -4,6 +4,7 @@ import React, { useState } from "react";
 import { Post } from "@prisma/client";
 import { fetchMorePosts } from "@/app/lib/fetches";
 import { useTranslations } from "next-intl";
+import PostDisplayed from "./posts-displayed";
 
 type ShowMoreButtonProps = {
   initialPosts: Post[];
@@ -11,7 +12,7 @@ type ShowMoreButtonProps = {
   locale: "en" | "es";
 };
 
-export default function ShowMoreButton({
+export default function ShowMoreLogic({
   initialPosts,
   totalPosts,
   locale,
@@ -28,13 +29,13 @@ export default function ShowMoreButton({
     setLoading(true);
     try {
       const { posts: newPosts, totalPosts: updatedTotalPosts } =
-        await fetchMorePosts(posts.length, 9, locale);
+        await fetchMorePosts(posts.length + 6, 9, locale);
 
       if (newPosts.length > 0) {
         setPosts((prevPosts) => [...prevPosts, ...newPosts]);
       }
 
-      if (posts.length + newPosts.length >= updatedTotalPosts) {
+      if (posts.length + 6 + newPosts.length >= updatedTotalPosts) {
         setAllPostsLoaded(true);
       }
     } catch (error) {
@@ -45,16 +46,17 @@ export default function ShowMoreButton({
   };
 
   return (
-    <>
+    <div className="flex flex-col">
+      <PostDisplayed initialPosts={posts} />
       {!allPostsLoaded && (
         <button
-          className="border-solid border-2 border-black text-black dark:border-white dark:text-white font-bold py-3 px-8 mb-10 rounded hover:bg-black dark:hover:bg-white hover:text-white dark:hover:text-black duration-200 transition-colors"
+          className="border-solid border-2 w-60 self-center my-8 border-black text-black dark:border-white dark:text-white font-bold py-3 px-8 mb-10 rounded hover:bg-black dark:hover:bg-white hover:text-white dark:hover:text-black duration-200 transition-colors"
           onClick={handleShowMore}
           disabled={loading}
         >
           {loading ? t("loading") : t("showMore")}
         </button>
       )}
-    </>
+    </div>
   );
 }
